requirements.txt
django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
corsheaders==4.0.0
python-decouple==3.8
PyJWT==2.8.0
djangorestframework-simplejwt==5.2.2


